tests/payment-tests.html

<!DOCTYPE html>
<html>
<head>
  <meta charset='utf-8'/>
  <title>Web Payments Tests</title>
  <link rel='stylesheet' href='../resources/testharness.css' media='all'/>
  <script src="../lib/paymentrequest.js"></script>
</head>
<body>
    <div><a href="../index.html">Index</a></div>
    <h1>Description</h1>
    <p>
      This test checks that the <code>PaymentRequest</code> runs correctly.
    </p>
    <div id='log'></div>
    <script src='../resources/testharness.js'></script>
    <script src='../resources/testharnessreport.js'></script>
    <script>
      window.__walletUrl = "payment-tests-wallet.html";

      test(function () {
          assert_true(undefined !== PaymentRequest, "PaymentRequest exists");
      }, "PaymentRequest is defined");

      /////////////////////
      // Constructor tests
      /////////////////////
      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest();
          }, "TypeError expected because methodData is missing");
      }, "PaymentRequest constructor validates that methodData argument is not optional");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest("string");
          }, "TypeError expected because methodData should be an array not a string");
      }, "PaymentRequest constructor validates that methodData is an array");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([]);
          }, "TypeError expected because methodData cannot be an empty array");
      }, "PaymentRequest constructor validates that methodData is not an empty array");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest(['method'],{ total:{}, displayItems:[{}] });
          }, "TypeError expected because methodData argument is not an array of objects");
      }, "PaymentRequest constructor validates that methodData argument is an array of objects");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{}],{ total:{}, displayItems:[{}] });
          }, "TypeError expected because methodData argument is missing supportedMethods");
      }, "PaymentRequest constructor validates that methodData argument has supportedMethods");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:[]}],{ total:{}, displayItems:[{}] });
          }, "TypeError expected because methodData argument has empty supportedMethods");
      }, "PaymentRequest constructor validates that methodData argument has non-empty supportedMethods");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:['method']}]);
          }, "TypeError expected because PaymentDetails argument is missing");
      }, "PaymentRequest constructor validates that PaymentDetails argument is not optional");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:['method']}], { });
          }, "TypeError expected because PaymentDetails.total argument is missing");
      }, "PaymentRequest constructor validates that PaymentDetails.total data is not optional");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:['method']}], { total: {} });
          }, "TypeError expected because total must have a non-negative amount argument is missing");
      }, "PaymentRequest constructor validates that total is non-negative");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:['method']}], { total: { amount: { value: "0.00" } }, displayItems: 'string' });
          }, "TypeError expected because PaymentDetails.displayItems argument must be an array");
      }, "PaymentRequest constructor validates that PaymentDetails.displayItems data is an array");

      test(function () {
          assert_throws(new TypeError(), function() {
            new PaymentRequest([{supportedMethods:['method']}], { total: { amount: { value: "0.00" } }, displayItems: [] });
          }, "TypeError expected because PaymentDetails.displayItems is of length 0");
      }, "PaymentRequest constructor validates that PaymentDetails.displayItems must have length > 0");

      test(function () {
      new PaymentRequest([{supportedMethods:['method']}], { total: { amount: { value: "0.00" } }, displayItems:[{}] });
      }, "PaymentRequest successfully constructed with valid arguments");

      /////////////////////
      // show() tests
      /////////////////////
      test(function () {
          assert_throws({name:"InvalidStateError"}, function() {
            var request = new PaymentRequest([{supportedMethods:['method']}], { total: { amount: { value: "0.00" } }, displayItems:[{}] });
            request.show();
            request.show();
          }, "InvalidStateError expected when calling show() twice");
      }, "Calling show() twice throws InvalidStateError");

      (function() {
        var test = async_test("show() with no accepted payment methods rejects promise with NotSupportedError");

        var request = new PaymentRequest(
            [{supportedMethods:['notaccepted']}],
            { total: { amount: { value: "0.00" } }, displayItems:[{}] }
          );

        request.show()
          .then(test.step_func(function(response) {
            assert_unreached("Promise should have been rejected with NotSupportedError but resolved instead.");
          }))
          .catch(test.step_func(function(e) {
            assert_equals(e.name,"NotSupportedError","Promise should have been rejected with NotSupportedError");
            assert_equals(request._debugOnlyState,"closed","After rejection, state should be closed");
            test.done();
        }));
      })();

      (function() {
        var test = async_test("show() with one accepted payment methods resolves promise with that method");

        var request = new PaymentRequest(
          [{supportedMethods:['accepted']}],
          { total: { amount: { value: "0.00" } }, displayItems:[{}] }
        );

        request.show()
          .then(test.step_func(function(response) {
            assert_equals(response.methodName,"accepted","Promise should have been resolved with 'accepted' method");
            assert_equals(request._debugOnlyState,"closed","State of request should be closed");
            response.complete(true);
            assert_equals(request._debugOnlyState,"closed","State of request should be closed after complete()");
            assert_throws({name:"InvalidStateError"}, function() {
              response.complete(false);
            },"complete() should throw when if called a second time");
            test.done();
          }))
          .catch(test.step_func(function(e) {
            assert_unreached("Promise should not have been rejected. " + e.name);
          }));
      })();

      (function() {
        var test = async_test("payment request with requestShipping resolves successfully");
        var addresschange = false, optionchange = false;

        var request = new PaymentRequest(
          [{supportedMethods:['accepted']}],
          { total: { amount: { value: "0.00" } }, displayItems:[{}] },
          { requestShipping:true }
        );
        request.onshippingaddresschange = function(e) {
          addresschange = true;
          e.updateWith(new Promise(function(resolve,reject) {
            resolve({shippingOptions:[]});
          }));
        };
        request.onshippingoptionchange = function(e) {
          optionchange = true;
          e.updateWith(new Promise(function(resolve,reject) {
            resolve({});
          }));
        };

        request.show()
          .then(test.step_func(function(response) {
            assert_equals(response.methodName,"accepted","Promise should have been resolved with 'accepted' method");
            assert_equals(request._debugOnlyState,"closed","State of request should be accepted");
            response.complete(true);
            assert_equals(request._debugOnlyState,"closed","State of request should be closed after complete()");
            assert_throws({name:"InvalidStateError"}, function() {
              response.complete(false);
            },"complete() should throw if called a second time");
            assert_true(addresschange,"onshippingaddresschange should have been called");
            assert_true(optionchange,"onshippingoptionchange should have been called");
            test.done();
          }))
          .catch(test.step_func(function(e) {
            assert_unreached("Promise should not have been rejected. " + e.name);
          }));
      })();

      /////////////////////
      // abort() tests
      /////////////////////
      test(function () {
          assert_throws({name:"InvalidStateError"}, function() {
            var request = new PaymentRequest(
              [{supportedMethods:['accepted']}],
              { total: { amount: { value: "0.00" } }, displayItems:[{}] }
            );
            request.abort();
          }, "InvalidStateError expected when calling abort() in created state");
      }, "Calling abort() in created state throws InvalidStateError");

    </script>
</body>
</html>
